moving compontnt
