+++ src/app/security/components/edit-user/index.ts
@@ -1,0 +1,1 @@
+export * from "./edit-user.component";
