+++ src/app/security/components/user-grid/index.ts
@@ -1,2 +1,1 @@
 export * from "./user-grid.component";
-export * from "./edit-user.component";
